requirements.txt
# At some point repo should be reorganised, currently every part is mushed together in one big folder
black
flask 
pyserial
flask-socketio


# dev
black